Replace text input with custom view in Restore screen
