+++ app/src/main/java/io/horizontalsystems/bankwallet/ui/extensions/InputTextView.kt
@@ -2,6 +2,8 @@ package io.horizontalsystems.bankwallet.ui.extensions
 
 import android.content.Context
 import android.support.constraint.ConstraintLayout
+import android.text.Editable
+import android.text.TextWatcher
 import android.util.AttributeSet
 import io.horizontalsystems.bankwallet.R
 import kotlinx.android.synthetic.main.view_text_input.view.*
@@ -25,6 +27,17 @@ class InputTextView : ConstraintLayout {
 
     fun getEnteredText(): String? {
         return inputEditText.text?.toString()
+    }
+
+    fun bindTextChangeListener(onTextChanged: ((String) -> Unit)) {
+        inputEditText.addTextChangedListener(object: TextWatcher {
+            override fun afterTextChanged(s: Editable) {
+                onTextChanged.invoke(s.toString().trim())
+            }
+
+            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) = Unit
+            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) = Unit
+        })
     }
 
 }
