+++ app/src/main/res/layout/view_holder_input_word.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<EditText
+<app:io.horizontalsystems.bankwallet.ui.extensions.InputTextView
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/wordInput"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:inputType="text"
-    android:fontFamily="@font/noto_sans"
-    android:layout_marginBottom="8dp"
-    android:imeOptions="actionNext"/>
+    android:layout_margin="4dp"
+    />
